// src/types/comment-types.ts
import {ObjectId} from "mongodb";

export type CommentType = {
    id: string
    content: string
    userId: string
    userLogin: string
    createdAt: string
}

export type CommentDbType = {
    _id: ObjectId
    id: string
    content: string
    userId: string
    userLogin: string
    createdAt: string
    postId: string
}